+++ pkg/dumpformat/testutils/parquet_writer.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package mydump
+package testutils
 
 import (
 	"context"
