+++ pkg/lightning/mydump/parquet_parser_test.go
@@ -34,6 +34,7 @@ import (
 	"github.com/apache/arrow-go/v18/parquet/file"
 	"github.com/apache/arrow-go/v18/parquet/metadata"
 	"github.com/apache/arrow-go/v18/parquet/schema"
+	"github.com/pingcap/tidb/pkg/dumpformat/testutils"
 	"github.com/pingcap/tidb/pkg/objstore"
 	"github.com/pingcap/tidb/pkg/objstore/objectio"
 	"github.com/pingcap/tidb/pkg/parser/mysql"
@@ -80,7 +81,7 @@ func newInt96FromUnixNanos(nanoseconds int64) parquet.Int96 {
 }
 
 func TestParquetParser(t *testing.T) {
-	pc := []ParquetColumn{
+	pc := []testutils.ParquetColumn{
 		{
 			Name:      "sS",
 			Type:      parquet.Types.ByteArray,
@@ -114,7 +115,7 @@ func TestParquetParser(t *testing.T) {
 
 	dir := t.TempDir()
 	name := "test123.parquet"
-	WriteParquetFile(dir, name, pc, 100)
+	testutils.WriteParquetFile(dir, name, pc, 100)
 
 	reader := newParquetParserForTest(context.Background(), t, dir, name, ParquetFileMeta{})
 
@@ -148,7 +149,7 @@ func TestParquetParser(t *testing.T) {
 	t.Run("unsupported_writer_option_is_rejected_before_create", func(t *testing.T) {
 		dir := t.TempDir()
 		fileName := "unsupported-option.parquet"
-		err := WriteParquetFile(dir, fileName, pc, 1, struct{}{})
+		err := testutils.WriteParquetFile(dir, fileName, pc, 1, struct{}{})
 		require.ErrorContains(t, err, "unsupported parquet writer option type")
 
 		_, statErr := os.Stat(filepath.Join(dir, fileName))
@@ -157,7 +158,7 @@ func TestParquetParser(t *testing.T) {
 }
 
 func TestParquetParserMultipleRowGroup(t *testing.T) {
-	pc := []ParquetColumn{
+	pc := []testutils.ParquetColumn{
 		{
 			Name:      "v",
 			Type:      parquet.Types.Int64,
@@ -176,7 +177,7 @@ func TestParquetParserMultipleRowGroup(t *testing.T) {
 
 	dir := t.TempDir()
 	fileName := "multi-row-group.parquet"
-	err := WriteParquetFile(
+	err := testutils.WriteParquetFile(
 		dir,
 		fileName,
 		pc,
@@ -198,7 +199,7 @@ func TestParquetParserMultipleRowGroup(t *testing.T) {
 
 func TestParquetVariousTypes(t *testing.T) {
 	t.Run("timestamp_and_decimal", func(t *testing.T) {
-		pc := []ParquetColumn{
+		pc := []testutils.ParquetColumn{
 			{
 				Name:      "date",
 				Type:      parquet.Types.Int32,
@@ -292,7 +293,7 @@ func TestParquetVariousTypes(t *testing.T) {
 
 		dir := t.TempDir()
 		name := "test123.parquet"
-		WriteParquetFile(dir, name, pc, 1)
+		testutils.WriteParquetFile(dir, name, pc, 1)
 
 		reader := newParquetParserForTest(context.Background(), t, dir, name, ParquetFileMeta{Loc: time.UTC})
 
@@ -341,7 +342,7 @@ func TestParquetVariousTypes(t *testing.T) {
 
 		const timestampMillis = int64(1603963672356)
 		const timestampMicros = int64(1603963672356956)
-		pc := []ParquetColumn{
+		pc := []testutils.ParquetColumn{
 			{
 				Name:    "timestamp_millis_instant",
 				Type:    parquet.Types.Int64,
@@ -378,7 +379,7 @@ func TestParquetVariousTypes(t *testing.T) {
 
 		dir := t.TempDir()
 		name := "logical-timestamps.parquet"
-		require.NoError(t, WriteParquetFile(dir, name, pc, 1))
+		require.NoError(t, testutils.WriteParquetFile(dir, name, pc, 1))
 
 		reader := newParquetParserForTest(context.Background(), t, dir, name, ParquetFileMeta{Loc: asiaShanghai})
 		require.Equal(t, schema.ConvertedTypes.TimestampMillis, reader.colTypes[0].converted)
@@ -412,7 +413,7 @@ func TestParquetVariousTypes(t *testing.T) {
 
 		const timeMillis = int32(123456)
 		const timeMicros = int64(123456789)
-		pc := []ParquetColumn{
+		pc := []testutils.ParquetColumn{
 			{
 				Name:    "time_millis_local",
 				Type:    parquet.Types.Int32,
@@ -433,7 +434,7 @@ func TestParquetVariousTypes(t *testing.T) {
 
 		dir := t.TempDir()
 		name := "logical-time-local.parquet"
-		require.NoError(t, WriteParquetFile(dir, name, pc, 1))
+		require.NoError(t, testutils.WriteParquetFile(dir, name, pc, 1))
 
 		reader := newParquetParserForTest(context.Background(), t, dir, name, ParquetFileMeta{Loc: asiaShanghai})
 		require.Equal(t, schema.ConvertedTypes.TimeMillis, reader.colTypes[0].converted)
@@ -456,7 +457,7 @@ func TestParquetVariousTypes(t *testing.T) {
 	})
 
 	t.Run("unsupported_logical_timestamp_nanos", func(t *testing.T) {
-		pc := []ParquetColumn{
+		pc := []testutils.ParquetColumn{
 			{
 				Name:    "timestamp_nanos",
 				Type:    parquet.Types.Int64,
@@ -469,7 +470,7 @@ func TestParquetVariousTypes(t *testing.T) {
 
 		dir := t.TempDir()
 		name := "logical-timestamp-nanos.parquet"
-		require.NoError(t, WriteParquetFile(dir, name, pc, 1))
+		require.NoError(t, testutils.WriteParquetFile(dir, name, pc, 1))
 
 		store, err := objstore.NewLocalStorage(dir)
 		require.NoError(t, err)
@@ -482,7 +483,7 @@ func TestParquetVariousTypes(t *testing.T) {
 
 	t.Run("int96_rounds_sub_microsecond_precision", func(t *testing.T) {
 		input := time.Date(2020, 10, 29, 9, 27, 52, 999999500, time.UTC)
-		pc := []ParquetColumn{
+		pc := []testutils.ParquetColumn{
 			{
 				Name:      "timestampint96",
 				Type:      parquet.Types.Int96,
@@ -495,7 +496,7 @@ func TestParquetVariousTypes(t *testing.T) {
 
 		dir := t.TempDir()
 		name := "int96-rounds-sub-microsecond.parquet"
-		require.NoError(t, WriteParquetFile(dir, name, pc, 1))
+		require.NoError(t, testutils.WriteParquetFile(dir, name, pc, 1))
 
 		reader := newParquetParserForTest(context.Background(), t, dir, name, ParquetFileMeta{Loc: time.UTC})
 		require.Equal(t, "", reader.colTypes[0].sparkRebaseMicros.timeZoneID)
@@ -563,7 +564,7 @@ func TestParquetVariousTypes(t *testing.T) {
 
 		for _, tc := range testCases {
 			t.Run(tc.name, func(t *testing.T) {
-				pc := []ParquetColumn{
+				pc := []testutils.ParquetColumn{
 					{
 						Name:      "date",
 						Type:      parquet.Types.Int32,
@@ -601,7 +602,7 @@ func TestParquetVariousTypes(t *testing.T) {
 				dir := t.TempDir()
 				name := "legacy-timestamp.parquet"
 				require.NoError(t,
-					WriteParquetFile(
+					testutils.WriteParquetFile(
 						dir, name, pc, 1,
 						parquet.WithCreatedBy("parquet-mr version 1.10.1"),
 						file.WithWriteMetadata(tc.meta),
@@ -747,7 +748,7 @@ func TestParquetVariousTypes(t *testing.T) {
 				require.NoError(t, legacyMeta.Append("org.apache.spark.timeZone", tc.timeZoneID))
 
 				legacyMicros := time.Date(1400, 3, 10, 0, 0, 0, 0, time.UTC).UnixMicro()
-				pc := []ParquetColumn{
+				pc := []testutils.ParquetColumn{
 					{
 						Name:      "timestampmillis",
 						Type:      parquet.Types.Int64,
@@ -777,7 +778,7 @@ func TestParquetVariousTypes(t *testing.T) {
 				dir := t.TempDir()
 				name := "legacy-timestamp-zone-table.parquet"
 				require.NoError(t,
-					WriteParquetFile(
+					testutils.WriteParquetFile(
 						dir, name, pc, 1,
 						parquet.WithCreatedBy("parquet-mr version 1.10.1"),
 						file.WithWriteMetadata(legacyMeta),
@@ -812,7 +813,7 @@ func TestParquetVariousTypes(t *testing.T) {
 		require.NoError(t, legacyMeta.Append("org.apache.spark.timeZone", "Unknown/Zone"))
 
 		legacyMicros := time.Date(1400, 3, 10, 0, 0, 0, 0, time.UTC).UnixMicro()
-		pc := []ParquetColumn{
+		pc := []testutils.ParquetColumn{
 			{
 				Name:      "timestampmicros",
 				Type:      parquet.Types.Int64,
@@ -826,7 +827,7 @@ func TestParquetVariousTypes(t *testing.T) {
 		dir := t.TempDir()
 		name := "legacy-timestamp-unknown-zone.parquet"
 		require.NoError(t,
-			WriteParquetFile(
+			testutils.WriteParquetFile(
 				dir, name, pc, 1,
 				parquet.WithCreatedBy("parquet-mr version 1.10.1"),
 				file.WithWriteMetadata(legacyMeta),
@@ -854,7 +855,7 @@ func TestParquetVariousTypes(t *testing.T) {
 		legacyMeta := metadata.NewKeyValueMetadata()
 		require.NoError(t, legacyMeta.Append("org.apache.spark.version", "2.4.8"))
 
-		pc := []ParquetColumn{
+		pc := []testutils.ParquetColumn{
 			{
 				Name:      "timestampmicros",
 				Type:      parquet.Types.Int64,
@@ -868,7 +869,7 @@ func TestParquetVariousTypes(t *testing.T) {
 		dir := t.TempDir()
 		name := "legacy-timestamp-system-zone.parquet"
 		require.NoError(t,
-			WriteParquetFile(
+			testutils.WriteParquetFile(
 				dir, name, pc, 1,
 				parquet.WithCreatedBy("parquet-mr version 1.10.1"),
 				file.WithWriteMetadata(legacyMeta),
@@ -913,7 +914,7 @@ func TestParquetVariousTypes(t *testing.T) {
 	})
 
 	t.Run("decimal_with_nulls", func(t *testing.T) {
-		pc := []ParquetColumn{
+		pc := []testutils.ParquetColumn{
 			{
 				Name:      "decimal1",
 				Type:      parquet.Types.Int32,
@@ -943,7 +944,7 @@ func TestParquetVariousTypes(t *testing.T) {
 
 		dir := t.TempDir()
 		fileName := "test.02.parquet"
-		WriteParquetFile(dir, fileName, pc, 7)
+		testutils.WriteParquetFile(dir, fileName, pc, 7)
 
 		reader := newParquetParserForTest(context.Background(), t, dir, fileName, ParquetFileMeta{})
 
@@ -965,7 +966,7 @@ func TestParquetVariousTypes(t *testing.T) {
 	})
 
 	t.Run("boolean", func(t *testing.T) {
-		pc := []ParquetColumn{
+		pc := []testutils.ParquetColumn{
 			{
 				Name:      "bool_val",
 				Type:      parquet.Types.Boolean,
@@ -978,7 +979,7 @@ func TestParquetVariousTypes(t *testing.T) {
 
 		dir := t.TempDir()
 		fileName := "test.bool.parquet"
-		WriteParquetFile(dir, fileName, pc, 2)
+		testutils.WriteParquetFile(dir, fileName, pc, 2)
 
 		reader := newParquetParserForTest(context.Background(), t, dir, fileName, ParquetFileMeta{})
 
@@ -1134,7 +1135,7 @@ func TestBasicReadFile(t *testing.T) {
 	rowCnt := 111
 	generated := make([]parquet.ByteArray, 0, rowCnt)
 
-	pc := []ParquetColumn{
+	pc := []testutils.ParquetColumn{
 		{
 			Name:      "s",
 			Type:      parquet.Types.ByteArray,
@@ -1157,7 +1158,7 @@ func TestBasicReadFile(t *testing.T) {
 	fileName := "test123.parquet"
 	// Genearte small file with multiple pages.
 	// The number of rows in each page is not multiple of batch size.
-	WriteParquetFile(dir, fileName, pc, rowCnt,
+	testutils.WriteParquetFile(dir, fileName, pc, rowCnt,
 		parquet.WithDataPageSize(512),
 		parquet.WithBatchSize(20),
 		parquet.WithCompressionFor("s", compress.Codecs.Uncompressed),
@@ -1187,7 +1188,7 @@ func TestParquetParserWrapper(t *testing.T) {
 	dir := t.TempDir()
 	fileName := "fixed-len-byte-array.parquet"
 
-	pc := []ParquetColumn{
+	pc := []testutils.ParquetColumn{
 		{
 			Name:      "fixed",
 			Type:      parquet.Types.FixedLenByteArray,
@@ -1218,7 +1219,7 @@ func TestParquetParserWrapper(t *testing.T) {
 			},
 		},
 	}
-	err := WriteParquetFile(
+	err := testutils.WriteParquetFile(
 		dir,
 		fileName,
 		pc,
@@ -1295,7 +1296,7 @@ func TestEstimateParquetReaderMemoryCtxLifetime(t *testing.T) {
 		},
 	)
 
-	pc := []ParquetColumn{
+	pc := []testutils.ParquetColumn{
 		{
 			Name:      "k",
 			Type:      parquet.Types.FixedLenByteArray,
@@ -1329,7 +1330,7 @@ func TestEstimateParquetReaderMemoryCtxLifetime(t *testing.T) {
 
 	dir := t.TempDir()
 	fileName := "ctx-lifetime.parquet"
-	require.NoError(t, WriteParquetFile(dir, fileName, pc, 50,
+	require.NoError(t, testutils.WriteParquetFile(dir, fileName, pc, 50,
 		parquet.WithMaxRowGroupLength(10),
 	))
